% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_python_modules.R
\name{check_for_python_modules}
\alias{check_for_python_modules}
\title{Look for Missing Python Modules and Optionally Install Them into a Virtual Environment}
\usage{
check_for_python_modules(silent = FALSE, install = FALSE)
}
\arguments{
\item{silent}{Logical. If \code{TRUE}, do not print results or progress to console.}

\item{install}{Logical.  If \code{TRUE}, install \code{numpy}, \code{scipy}, or \code{nmrglue}
modules if missing.}
}
\value{
Invisibly, a named logical vector giving the success or not of finding
and possibly installing the modules of interest.
}
\description{
This function will check to see if the Python modules \code{numpy},
\code{scipy}, and \code{nmrglue} are installed, and optionally, install them into the
virtual environment \code{r-reticulate}.
}
\author{
Bryan A. Hanson, hanson@depauw.edu
}
