% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Bruker_1D_NMR.R
\name{import_Bruker_1D_NMR}
\alias{import_Bruker_1D_NMR}
\title{Import Bruker 1D NMR Data Sets}
\usage{
import_Bruker_1D_NMR(file)
}
\arguments{
\item{file}{Character.  The path to the \code{pdata} file.}
}
\value{
COMPLETE THIS!
}
\description{
This function will read a Bruker 1D NMR data set and return the pieces.
Currently only processed data can be read.
}
\author{
Bryan A. Hanson
}
