% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_session_info.R
\name{python_session_info}
\alias{python_session_info}
\title{Provide Information About the Python Configuration}
\usage{
python_session_info(silent = FALSE)
}
\arguments{
\item{silent}{Logical. If \code{TRUE}, do not print results or progress to console.}
}
\value{
Invisibly, a named logical vector giving the success or not of finding
and possibly installing the modules of interest.
}
\description{
Report on the Python configuration.  Similar in spirit and use to
\code{sessionInfo}.
}
\author{
Bryan A. Hanson, hanson@depauw.edu
}
