% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Bruker_1D_NMR.R
\name{read_Bruker_1D_NMR}
\alias{read_Bruker_1D_NMR}
\title{Read Bruker 1D NMR Data Sets}
\usage{
read_Bruker_1D_NMR(directory)
}
\arguments{
\item{directory}{Character.  The path to the \code{pdata} file.}
}
\value{
COMPLETE THIS!
}
\description{
This function will read a Bruker 1D NMR data set and return the pieces.
Currently only processed data can be read.
}
\examples{
data_dir <- system.file("extdata/bmse000233/nmr/set01/1H/pdata/1",
  package = "readBrukerNMR")
demo <- read_Bruker_1D_NMR(data_dir)
plot(demo$x, demo$y, type = "l",
  xlim = c(5.5, 3), ylim = c(0, 6e4)
  xlab = "ppm", ylab = "intensity")

}
\author{
Bryan A. Hanson
}
