% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrukerNMR-package.R
\docType{package}
\name{readBrukerNMR-package}
\alias{readBrukerNMR-package}
\alias{readBrukerNMR}
\title{Read Bruker NMR Data Sets}
\description{
Read in Bruker NMR (Nuclear Magnetic Resonance) data.  Currently only reads processed data.
Built around the 'Python' 'nmrglue' module by JJ Helmus.
}
\author{
Bryan A. Hanson

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}
