% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBrukerNMR-package.R
\docType{package}
\name{readBrukerNMR-package}
\alias{readBrukerNMR-package}
\alias{readBrukerNMR}
\title{Read Bruker NMR Data Sets}
\description{
Read in Bruker NMR (Nuclear Magnetic Resonance) data.  Currently only reads processed data.
Built around the Python \href{https://github.com/jjhelmus/nmrglue/}{'nmrglue'} module by JJ Helmus.
Includes a data set from \href{https://bmrb.io/}{BMRB} with permission for demonstration and testing.
The included data are 1H and 13C NMR spectra of the disaccharide melibiose.  These spectra are
a portion of the
\href{https://bmrb.io/metabolomics/mol_summary/show_data.php?id=bmse000233}{bmse000233 data set}
hosted at the
\href{https://dx.doi.org/10.1093/nar/gkm957}{Biological Magnetic Resonance Data Bank} (used with permission).
}
\author{
Bryan A. Hanson

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}
